fixed more problems with glibc stage 3
bison is now a dependency for glibc, so had to move the bison build
section before the glibc section.

also, the use of a build directory is required by the glibc configure
script, so added directions to make one and change into it. also
reverted the doubled dot (..) on the configure script.

fixed typo.